Create Dockerfile
This Dockerfile is used to create a Docker image that runs a PHP web application with Apache web server. The base image used is php:7.2-apache, which includes the Apache web server and the PHP 7.2 runtime. The Dockerfile copies the hello.php file into the Apache web server's document root directory (/var/www/html/), which is where Apache serves the default index file from. The EXPOSE instruction is used to specify that the container will listen on port 8080 for incoming connections. When the container is run, it will start the Apache web server and serve the hello.php file when the user visits the website.